requirements.txt
fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.24.0,<0.30.0
pydantic>=2.5.0,<3.0.0
pandas>=2.2.0,<3.0.0
xlsxwriter>=3.1.0,<4.0.0
python-dateutil>=2.8.2
pytz>=2023.3
python-dotenv>=1.0.0
PyJWT>=2.8.0,<3.0.0
psycopg2-binary>=2.9.9
